Fixing favorites bug
